wip: sound manager tab [macata #94]
